Support for pre-London blocks missing a base fee during validation
Set ``base_fee_per_gas`` value to ``None`` during validation if the param is missing (pre-London blocks). Pop this value back out during block normalization. This returns a block without a ``base_fee_per_gas`` for pre-London blocks which seems to be the default behavior in other clients as well.